LCORE-636: split config Customization unit tests
